feat: Refactor string formatting for improved readability and consistency
**Key Updates**:
- Updated string formatting across multiple files to use the more concise `{variable}` syntax instead of `format!("...{}", variable)`.
- Enhanced code clarity and maintainability by standardizing error messages and logging outputs.

**Outcome**:
- Improved overall code readability and reduced verbosity in string handling, contributing to a cleaner codebase.

**Conclusion**: ✅ Successfully refactored string formatting, enhancing code quality and maintainability!

+++ crates/agent-mem-intelligence/src/processing/adaptive.rs
@@ -238,7 +238,7 @@ impl AdaptiveMemoryManager {
 
         // Check for deletion conditions
         if age > self.thresholds.delete_age_threshold
-            || (memory.score.unwrap_or(0.5) < self.thresholds.min_importance as f32
+            || (memory.score.unwrap_or(0.5) < self.thresholds.min_importance
                 && self.get_access_count(memory) < self.thresholds.min_access_count)
         {
             return Ok(LifecycleAction::Delete);
